Re-arrange pad2d/pad3d to pad similar to nabla, re-arrange module adding, add complex-valued processing.
