+++ python/setup.py.in
@@ -86,6 +86,31 @@ class BuildExtension(build_ext):
             ext.extra_compile_args = opts
         build_ext.build_extensions(self)
 
+EXT_MODULES = []
+
+def add_module(opname, inc_dir=[]):
+    EXT_MODULES.append(
+        Extension(f"py_{opname}_operator", [
+            os.path.join("${CMAKE_CURRENT_SOURCE_DIR}", f"py_{opname}_operator.cpp"),
+            ],
+            include_dirs=["${OPTOX_INCLUDE_DIR}",
+                "${CUDA_INCLUDE_DIRS}",
+                "${CUDA_SDK_ROOT_DIR}",
+                get_numpy_include(),
+                get_pybind_include(),
+            ] + inc_dir,
+            library_dirs=["${OPTOX_LIBRARY_DIR}",
+                os.path.join("${CUDA_ROOT_DIR}", "lib64"),],
+            runtime_library_dirs=["${OPTOX_LIBRARY_DIR}",
+                os.path.join("${CUDA_ROOT_DIR}", "lib64"),],
+            libraries=["optox", "cudart"]
+        ),
+    )
+
+add_module("nabla")
+add_module("pad")
+
+
 setup(
     name='optopy',
     version='0.2.0',
@@ -95,22 +120,7 @@ setup(
     package_dir={"optopy": os.path.join("${CMAKE_CURRENT_SOURCE_DIR}", "optopy"),},
     include_dirs=[],
     ext_package="_ext",
-    ext_modules=[
-        Extension("py_nabla_operator", [
-            os.path.join("${CMAKE_CURRENT_SOURCE_DIR}", "py_nabla_operator.cpp"),
-            ],
-            include_dirs=["${OPTOX_INCLUDE_DIR}",
-                "${CUDA_INCLUDE_DIRS}",
-                get_numpy_include(),
-                get_pybind_include(),
-            ],
-            library_dirs=["${OPTOX_LIBRARY_DIR}",
-                os.path.join("${CUDA_TOOLKIT_ROOT_DIR}", "lib64"),],
-            runtime_library_dirs=["${OPTOX_LIBRARY_DIR}",
-                os.path.join("${CUDA_TOOLKIT_ROOT_DIR}", "lib64"),],
-            libraries=["optox", "cudart"]
-        ),
-    ],
+    ext_modules=EXT_MODULES,
     cmdclass={
         'build_ext': BuildExtension
     },
