+++ pytorch/setup.py.in
@@ -2,6 +2,27 @@ import os
 from setuptools import setup
 from torch.utils.cpp_extension import BuildExtension, CppExtension, CUDAExtension
 
+EXT_MODULES = []
+
+def add_module(opname, inc_dir=[]):
+    EXT_MODULES.append(
+        CUDAExtension(f"th_{opname}_operator", [
+            os.path.join("${CMAKE_CURRENT_SOURCE_DIR}", f"th_{opname}_operator.cpp"),
+            ],
+            include_dirs=["${OPTOX_INCLUDE_DIR}", "${CUDA_SDK_ROOT_DIR}"] + inc_dir,
+            library_dirs=["${OPTOX_LIBRARY_DIR}",],
+            runtime_library_dirs=["${OPTOX_LIBRARY_DIR}",],
+            libraries=["optox", ]
+        ),
+    )
+
+add_module("act")
+add_module("nabla")
+add_module("demosaicing")
+add_module("rot")
+add_module("warp")
+add_module("pad")
+
 setup(
     name='optoth',
     version='0.2.0',
@@ -13,64 +34,7 @@ setup(
                  "optoth.activations": os.path.join("${CMAKE_CURRENT_SOURCE_DIR}", "optoth", "activations")},
     include_dirs=[],
     ext_package="_ext",
-    ext_modules=[
-        CUDAExtension("th_act_operators", [
-            os.path.join("${CMAKE_CURRENT_SOURCE_DIR}", "th_act_operator.cpp"),
-            ],
-            include_dirs=["${OPTOX_INCLUDE_DIR}",],
-            library_dirs=["${OPTOX_LIBRARY_DIR}",],
-            runtime_library_dirs=["${OPTOX_LIBRARY_DIR}",],
-            libraries=["optox", ]
-        ),
-        CUDAExtension("th_nabla_operator", [
-            os.path.join("${CMAKE_CURRENT_SOURCE_DIR}", "th_nabla_operator.cpp"),
-            ],
-            include_dirs=["${OPTOX_INCLUDE_DIR}",],
-            library_dirs=["${OPTOX_LIBRARY_DIR}",],
-            runtime_library_dirs=["${OPTOX_LIBRARY_DIR}",],
-            libraries=["optox", ]
-        ),
-        CUDAExtension("th_demosaicing_operator", [
-            os.path.join("${CMAKE_CURRENT_SOURCE_DIR}", "th_demosaicing_operator.cpp"),
-            ],
-            include_dirs=["${OPTOX_INCLUDE_DIR}",],
-            library_dirs=["${OPTOX_LIBRARY_DIR}",],
-            runtime_library_dirs=["${OPTOX_LIBRARY_DIR}",],
-            libraries=["optox", ]
-        ),
-        CUDAExtension("th_rot_operator", [
-            os.path.join("${CMAKE_CURRENT_SOURCE_DIR}", "th_rot_operator.cpp"),
-            ],
-            include_dirs=["${OPTOX_INCLUDE_DIR}",],
-            library_dirs=["${OPTOX_LIBRARY_DIR}",],
-            runtime_library_dirs=["${OPTOX_LIBRARY_DIR}",],
-            libraries=["optox", ]
-        ),
-        CUDAExtension("th_warp_operator", [
-            os.path.join("${CMAKE_CURRENT_SOURCE_DIR}", "th_warp_operator.cpp"),
-            ],
-            include_dirs=["${OPTOX_INCLUDE_DIR}",],
-            library_dirs=["${OPTOX_LIBRARY_DIR}",],
-            runtime_library_dirs=["${OPTOX_LIBRARY_DIR}",],
-            libraries=["optox", ]
-        ),
-        CUDAExtension("th_pad2d_operator", [
-            os.path.join("${CMAKE_CURRENT_SOURCE_DIR}", "th_pad2d_operator.cpp"),
-            ],
-            include_dirs=["${OPTOX_INCLUDE_DIR}",],
-            library_dirs=["${OPTOX_LIBRARY_DIR}",],
-            runtime_library_dirs=["${OPTOX_LIBRARY_DIR}",],
-            libraries=["optox", ]
-        ),
-        CUDAExtension("th_pad3d_operator", [
-            os.path.join("${CMAKE_CURRENT_SOURCE_DIR}", "th_pad3d_operator.cpp"),
-            ],
-            include_dirs=["${OPTOX_INCLUDE_DIR}",],
-            library_dirs=["${OPTOX_LIBRARY_DIR}",],
-            runtime_library_dirs=["${OPTOX_LIBRARY_DIR}",],
-            libraries=["optox", ]
-        ),
-    ],
+    ext_modules=EXT_MODULES,
     cmdclass={
         'build_ext': BuildExtension
     },
