+++ src/CMakeLists.txt
@@ -22,8 +22,7 @@ set(OPERATOR_HEADER
     operators/nabla2_operator.h
     operators/rot_operator.h
     operators/warp_operator.h
-    operators/pad2d_operator.h
-    operators/pad3d_operator.h
+    operators/pad_operator.h
     operators/demosaicing_operator.h
 )
 
@@ -35,8 +34,7 @@ set(OPERATOR_SOURCE
     operators/nabla2_operator.cu
     operators/rot_operator.cu
     operators/warp_operator.cu
-    operators/pad2d_operator.cu
-    operators/pad3d_operator.cu
+    operators/pad_operator.cu
     operators/demosaicing_operator.cu
 )
